chore: refactor gooleapi module
